DiT/requirement.txt
pytorch
torchvision
timm
diffusers
accelerate